Robel

+++ app/app/Http/Controllers/API/ProductController.php
@@ -15,6 +15,7 @@ use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Storage;
 
 class ProductController extends ApiController
 {
@@ -25,7 +26,7 @@ class ProductController extends ApiController
      * @param \Illuminate\Http\Request $request
      * @return JsonResponse
      */
-    public function index(Request $request):JsonResponse
+    public function index(Request $request): JsonResponse
     {
 
         $company = $request->company;
@@ -119,7 +120,7 @@ class ProductController extends ApiController
      * @param int $id
      * @return JsonResponse
      */
-    public function editProduct($id):JsonResponse
+    public function editProduct($id): JsonResponse
     {
         $product = Product::with(['productVarients'])->where('id', $id)->first();
 
@@ -138,22 +139,29 @@ class ProductController extends ApiController
      * @param int $id
      * @return JsonResponse
      */
-    public function updateProduct(UpdateRequest $request, $id):JsonResponse
+    public function updateProduct(UpdateRequest $request, $id)
     {
+
         try {
-            $product = ProductProcess::update($request, $id);
+            $product =  Product::find($id);
+            if(!empty($product)){
 
-            $arrayofProductVarientId = ProductVarient::where('product_id', $id)->pluck('id')->toArray();
+                $product = ProductProcess::update($request, $id);
 
-            $deletableProductVarient = $this->updateProductVarient($request, $product, $arrayofProductVarientId);
+                $arrayofProductVarientId = ProductVarient::where('product_id', $id)->pluck('id')->toArray();
 
-            ProductVarient::whereIn('id', $deletableProductVarient)->delete();
+                $deletableProductVarient = $this->updateProductVarient($request, $product, $arrayofProductVarientId);
 
-            if (isset($product->productVarients)) {
-                $product->productVarients;
+                ProductVarient::whereIn('id', $deletableProductVarient)->delete();
+
+                if (isset($product->productVarients)) {
+                    $product->productVarients;
+                }
+
+                return $this->jsonResponse(false, "Product updated successfully", $product, $this->emptyArray, JsonResponse::HTTP_OK);
+            }else{
+                return $this->jsonResponse(false,$this->failed,['Product not found'], $this->emptyArray, JsonResponse::HTTP_NOT_FOUND);
             }
-
-            return $this->jsonResponse(false, "Product updated successfully", $product, $this->emptyArray, JsonResponse::HTTP_OK);
         } catch (\Exception $e) {
 
             return $this->jsonResponse(true, 'Failed to update product', $request->all(), [$e->getMessage()], JsonResponse::HTTP_INTERNAL_SERVER_ERROR);
@@ -167,7 +175,7 @@ class ProductController extends ApiController
      * @param int $id
      * @return JsonResponse
      */
-    public function productDetails(Request $request, $id):JsonResponse
+    public function productDetails(Request $request, $id): JsonResponse
     {
 
         $query = Product::with(['productVarients', 'company', 'reviews' => function ($query) {
@@ -218,7 +226,7 @@ class ProductController extends ApiController
      * @param $productId
      * @return JsonResponse
      */
-    public function companyProductDetails($companyId, $productId):JsonResponse
+    public function companyProductDetails($companyId, $productId): JsonResponse
     {
         $product = Product::with(['productVarients'])
             ->where('company_id', $companyId)
@@ -246,12 +254,29 @@ class ProductController extends ApiController
 
         $product = Product::where('user_id', Auth::id())->where('id', $id)->first();
 
+        //delete product images
+        if (isset($product->images)) {
+            $arrayofImages = json_decode($product->images);
+            $this->deleteFile("public", $arrayofImages);
+        }
+
+        //delete product varient images
+        if(isset($product->productVarients))
+        {
+            foreach ($product->productVarients as $proVarient)
+            {
+                $arrayofImages = json_decode($proVarient->images);
+                $this->deleteFile("public", $arrayofImages);
+            }
+        }
+
         if (!empty($product)) {
 
             $product->delete();
 
             return $this->jsonResponse(false, 'Product deleted successfully', $product, $this->emptyArray, JsonResponse::HTTP_OK);
         } else {
+
             return $this->jsonResponse(true, $this->failed, $this->emptyArray, ['Product not found'], JsonResponse::HTTP_NOT_FOUND);
         }
     }
@@ -262,7 +287,7 @@ class ProductController extends ApiController
      * @param $arrayofProductVarientId
      * @return array
      */
-    protected function updateProductVarient($request, $product, $arrayofProductVarientId):array
+    protected function updateProductVarient($request, $product, $arrayofProductVarientId): array
     {
         if (isset($request->product_varients)) {
             foreach ($request->product_varients as $productVarient) {
